+++ NEWS.md
@@ -1,5 +1,7 @@
 # teal.code 0.4.1.9001
 
+* The `@code` field in the `qenv` class now holds `character`, not `expression`.
+
 # teal.code 0.4.1
 
 ### Miscellaneous